+++ src/cli/cli.go
@@ -18,6 +18,8 @@ func (c *CliApp) Initialize() {
 func (c *CliApp) Run() {
 	if c.SelectedPing() {
 		c.Ping()
+	} else if c.SelectedConfigure() {
+		c.Configure()
 	} else if c.SelectedDefault() {
 		fmt.Println("i am the master of war!")
 	} else {