+++ src/cli/configure.go
@@ -1,0 +1,11 @@
+package cli
+
+import (
+	"fmt"
+)
+
+func (c *CliApp) Configure() {
+	// ask for token
+	// delete token in spekkiofile if exists
+	// write token to spekkiofile
+}